test(ui): split sidebar PR state coverage
